Started splitting up NewGrid into separate components, experimenting with removing layoutTransition.
