+++ src/components/NewGrid.js
@@ -1,4 +1,4 @@
-import React, { memo, useEffect, useState, useRef } from 'react';
+import React, { memo, useEffect, useState, useRef, useLayoutEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import ReactRouterPropTypes from 'react-router-prop-types';
@@ -7,6 +7,7 @@ import { FiInstagram } from 'react-icons/fi';
 import { motion, useInvertedScale, useMotionValue, useAnimation } from 'framer-motion';
 import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
 import { Loading } from './Loading';
+import { Image } from './Image';
 
 // Separate into individual components
 const StyledGrid = styled.div`
@@ -83,29 +84,6 @@ const StyledGrid = styled.div`
     right: 0;
     bottom: 0;
   }
-  .post-image-container {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-    transform: translateZ(0);
-    object-fit: none;
-    object-position: center center;
-    background: rgba(0, 0, 0, 0.05);
-  }
-  .post-image-container img {
-    display: block;
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    opacity: 0;
-    /* object-fit: cover; */
-    /* object-fit: none;
-    object-position: 50% 50%; */
-  }
-  .post-image.open {
-    width: 100%;
-  }
   .overlay {
     z-index: 1;
     position: fixed;
@@ -134,11 +112,9 @@ const StyledGrid = styled.div`
     font-family: 'Open Sans', sans-serif;
     color: #032d4d;
     font-weight: 400;
-    /* color: #121212; */
     background: #fff;
     display: none;
     font-weight: 500;
-    /* max-width: 9%; */
     opacity: 0;
     p {
       margin: 0;
@@ -147,14 +123,10 @@ const StyledGrid = styled.div`
       margin: 0 auto;
     }
     @media only screen and (max-width: 700px) {
-      /* .caption-container { */
       font-size: 0.8em;
-      /* } */
     }
     @media only screen and (max-width: 500px) {
-      /* .caption-container { */
       font-size: 0.9em;
-      /* } */
     }
   }
   .caption-container .open {
@@ -164,23 +136,16 @@ const StyledGrid = styled.div`
     z-index: 2000 !important;
     position: relative;
     font-size: 1em;
-    /* padding-top: 1em; */
     opacity: 0.7;
-    /* color: #121212; */
     /* allows us to click through overlay to actually use link */
     pointer-events: auto;
     display: block;
   }
 
   .links {
-    /* display: inline; */
-
-    /* padding-right: 5px; */
     a,
     a:link {
       font-size: 0.8em;
-      /* display: inline; */
-      /* float: right; */
       bottom: 0;
       margin-right: 0.8rem;
       margin-bottom: 0.45rem;
@@ -190,13 +155,12 @@ const StyledGrid = styled.div`
       position: absolute;
       transition: color 0.2s ease-out;
       @media only screen and (min-width: 1000px) {
+        /* I think this is only scoped to this element? */
         font-size: 1em;
-
         margin-right: 1rem;
       }
     }
     a:hover {
-      /* color: #121212; */
       color: hsla(205.9, 85.3%, 40%, 1);
     }
   }
@@ -227,7 +191,7 @@ const StyledGrid = styled.div`
 
   @media only screen and (max-width: 750px) {
     .open .post-content {
-      max-width: 100vw;
+      max-width: 90vw;
     }
     padding: 0 25px;
     .post {
@@ -284,9 +248,6 @@ export function NewGrid({ posts, match, history }) {
 
   useEffect(() => {
     if (posts.length !== 0) {
-      // setTimeout(() => {
-      //   setLoading(false);
-      // }, 500);
       setLoading(false);
     }
   }, [posts]);
@@ -305,7 +266,7 @@ export function NewGrid({ posts, match, history }) {
   return (
     <StyledGrid>
       <div className="grid">
-        {loading && false ? (
+        {loading ? (
           <Loading />
         ) : (
           !!posts &&
@@ -325,167 +286,133 @@ export function NewGrid({ posts, match, history }) {
   );
 }
 
-// const Post = memo(
-const Post = ({ isSelected, post, maxHeight, history }) => {
-  const [fromGrid, setFromGrid] = useState(false);
-  const zIndex = useMotionValue(isSelected ? 2 : 0);
-  const postRef = useRef(null);
-  const containerRef = useRef(null);
+const Post = memo(
+  ({ isSelected, post, maxHeight, history }) => {
+    const [fromGrid, setFromGrid] = useState(false);
+    const zIndex = useMotionValue(isSelected ? 2 : 0);
+    const postRef = useRef(null);
+    const containerRef = useRef(null);
 
-  function checkZIndex() {
-    if (isSelected) {
-      zIndex.set(2);
-    } else if (!isSelected) {
-      zIndex.set(0);
-    }
-  }
-
-  // dismiss modal when escape is pressed
-  useEffect(() => {
-    const dismissModal = event => {
-      if (isSelected && event.key === 'Escape') {
-        history.push('/');
+    function checkZIndex() {
+      if (isSelected) {
+        zIndex.set(2);
+      } else if (!isSelected) {
+        zIndex.set(0);
       }
-    };
-
-    window.addEventListener('keydown', dismissModal);
-
-    return () => {
-      window.removeEventListener('keydown', dismissModal);
-    };
-  }, [isSelected, history]);
-
-  // when modal is dismissed, make sure scroll pos is in sync
-  // when visiting an item near end of list directly, when modal dismissed
-  // scroll pos was top of list
-  useEffect(() => {
-    const scrollToRef = ref =>
-      window.scrollTo({
-        top: ref.current.offsetTop - ref.current.offsetHeight,
-        behaviour: 'smooth',
-      });
-    if (isSelected && !fromGrid) {
-      scrollToRef(containerRef);
     }
-  }, [fromGrid, isSelected]);
 
-  return (
-    <div className="post" style={{ maxHeight }} ref={containerRef}>
-      <Overlay isSelected={isSelected} />
-      <div className={`post-content-container ${isSelected && 'open'}`}>
-        <motion.div
-          ref={postRef}
-          // without layout transition, zIndex doesn't update
-          layoutTransition={isSelected ? closeSpring : openSpring}
-          style={{ zIndex }}
-          className="post-content"
-          onUpdate={checkZIndex}
-          drag={isSelected && false}
-        >
-          <Image
-            id={post.id}
-            isSelected={isSelected}
-            src={post.src}
-            caption={post.caption}
-            width={post.width}
-            height={post.height}
+    // dismiss modal when escape is pressed
+    useEffect(() => {
+      const dismissModal = event => {
+        if (isSelected && event.key === 'Escape') {
+          history.push('/');
+        }
+      };
+
+      window.addEventListener('keydown', dismissModal);
+
+      return () => {
+        window.removeEventListener('keydown', dismissModal);
+      };
+    }, [isSelected, history]);
+
+    // when modal is dismissed, make sure scroll pos is in sync
+    // when visiting an item near end of list directly, when modal dismissed
+    // scroll pos was top of list
+    useEffect(() => {
+      const scrollToRef = ref =>
+        window.scrollTo({
+          top: ref.current.offsetTop - ref.current.offsetHeight,
+          behaviour: 'smooth',
+        });
+      if (isSelected && !fromGrid) {
+        scrollToRef(containerRef);
+      }
+    }, [fromGrid, isSelected]);
+
+    useEffect(() => {
+      if (isSelected) {
+        zIndex.set(2);
+      } else if (!isSelected) {
+        zIndex.set(0);
+      }
+    }, [zIndex, isSelected]);
+
+    return (
+      <div className="post" style={{ maxHeight }} ref={containerRef}>
+        <Overlay isSelected={isSelected} />
+        <div className={`post-content-container ${isSelected && 'open'}`}>
+          <motion.div
+            ref={postRef}
+            // without layout transition, zIndex doesn't update
+            transition={openSpring}
+            // positionTransition={isSelected ? closeSpring : openSpring}
+            style={{ zIndex }}
+            className="post-content"
+            onUpdate={checkZIndex}
+            // drag={isSelected}
+          >
+            <Image
+              id={post.id}
+              isSelected={isSelected}
+              src={post.src}
+              caption={post.caption}
+              width={post.width}
+              height={post.height}
+            />
+            <Caption caption={post.caption} isSelected={isSelected} id={post.id} link={post.link} />
+          </motion.div>
+        </div>
+
+        {!isSelected && (
+          <Link
+            to={{
+              pathname: `posts/${post.id}`,
+            }}
+            className="post-open-link"
+            onClick={() => setFromGrid(true)}
           />
-          <Caption caption={post.caption} isSelected={isSelected} id={post.id} link={post.link} />
-        </motion.div>
+        )}
       </div>
+    );
+  },
+  (prev, next) => prev.isSelected === next.isSelected,
+);
 
-      {!isSelected && (
-        <Link
-          to={{
-            pathname: `posts/${post.id}`,
-          }}
-          className="post-open-link"
-          onClick={() => setFromGrid(true)}
-        />
-      )}
-    </div>
-  );
-};
-// (prev, next) => prev.isSelected === next.isSelected,
-// );
-const ImagePlaceholder = styled.div`
-  background-image: radial-gradient(
-    to right,
-    rgba(0, 0, 0, 0.8) 0,
-    rgba(0, 0, 0, 0.17) 15%,
-    rgba(0, 0, 0, 0.1) 30%
-  );
-  background-size: 600px 100%;
-  /* height: 400px; */
-  position: static;
-  overflow: hidden;
-  /* -webkit-animation: placeholderShimmer 2s linear;
-  animation: placeholderShimmer 2s linear;
-  -webkit-animation-iteration-count: infinite;
-  animation-iteration-count: infinite; */
-  -webkit-transform: translateZ(0);
-  -moz-transform: translateZ(0);
-  -ms-transform: translateZ(0);
-  -o-transform: translateZ(0);
-  transform: translateZ(0);
+// function Image({ isSelected, id, src, height, width, caption }) {
+//   const controls = useAnimation();
+//   const [loaded, setLoaded] = useState(false);
 
-  --backgroundOffset: 600px;
+//   const inverted = useInvertedScale();
 
-  @-webkit-keyframes placeholderShimmer {
-    0% {
-      background-position: (-1 * var(--backgroundOffset)) 0;
-    }
+//   useEffect(() => {
+//     if (!loaded) return;
+//     controls.start({
+//       opacity: 1,
+//       transition: {
+//         duration: 0.5,
+//       },
+//     });
+//   }, [loaded]);
 
-    100% {
-      background-position: var(--backgroundOffset) 0;
-    }
-  }
-
-  @keyframes placeholderShimmer {
-    0% {
-      background-position: (-1 * var(--backgroundOffset)) 0;
-    }
-
-    100% {
-      background-position: var(--backgroundOffset) 0;
-    }
-  }
-`;
-
-function Image({ isSelected, id, src, height, width, caption }) {
-  const controls = useAnimation();
-  const [loaded, setLoaded] = useState(false);
-
-  const inverted = useInvertedScale();
-
-  useEffect(() => {
-    if (!loaded) return;
-    controls.start({
-      opacity: 1,
-      transition: {
-        duration: 0.5,
-      },
-    });
-  }, [loaded]);
-
-  return (
-    <motion.div className="post-image-container" style={{ ...inverted, originX: 0, originY: 0 }}>
-      {!loaded && false && <ImagePlaceholder style={{ height, width }} />}
-      <motion.img
-        key={`post-${id}`}
-        className={`post-image ${isSelected && 'open'}`}
-        src={src}
-        alt={caption}
-        transition={{ ...closeSpring, duration: 0.4 }}
-        animate={controls}
-        onLoad={() => setLoaded(true)}
-        style={{
-          display: `${loaded ? 'block' : 'none'}`,
-        }}
-      />
-    </motion.div>
-  );
-}
+//   return (
+//     <motion.div className="post-image-container" style={{ ...inverted, originX: 0, originY: 0 }}>
+//       {!loaded && false && <ImagePlaceholder style={{ height, width }} />}
+//       <motion.img
+//         key={`post-${id}`}
+//         className={`post-image ${isSelected && 'open'}`}
+//         src={src}
+//         alt={caption}
+//         transition={{ ...closeSpring, duration: 0.4 }}
+//         animate={controls}
+//         onLoad={() => setLoaded(true)}
+//         style={{
+//           display: `${loaded ? 'block' : 'none'}`,
+//         }}
+//       />
+//     </motion.div>
+//   );
+// }
 
 function Caption({ isSelected, caption, link }) {
   const opacity = isSelected ? 1 : 0;
